+++ src/main/java/com/fastcampus/reserve/infrestructure/product/ProductReaderImpl.java
@@ -1,10 +1,9 @@
 package com.fastcampus.reserve.infrestructure.product;
 
 
-import com.fastcampus.reserve.domain.dto.request.ProductListOptionDto;
+import com.fastcampus.reserve.domain.product.dto.request.ProductListOptionDto;
 import com.fastcampus.reserve.domain.product.Product;
 import com.fastcampus.reserve.domain.product.ProductReader;
-import java.time.LocalDate;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
