+++ src/main/java/com/fastcampus/reserve/interfaces/product/ProductDtoMapper.java
@@ -1,15 +1,11 @@
 package com.fastcampus.reserve.interfaces.product;
 
-import com.fastcampus.reserve.domain.dto.response.ProductDto;
-import com.fastcampus.reserve.domain.product.Product;
-import com.fastcampus.reserve.domain.product.ProductImage;
+import com.fastcampus.reserve.domain.product.dto.response.ProductDto;
 import com.fastcampus.reserve.interfaces.product.dto.response.ProductResponse;
 import java.util.ArrayList;
 import java.util.List;
 import org.mapstruct.InjectionStrategy;
 import org.mapstruct.Mapper;
-import org.mapstruct.Mapping;
-import org.mapstruct.Named;
 import org.mapstruct.ReportingPolicy;
 
 @Mapper(
